atualizacao

+++ models/CaminhaoCliente.php
@@ -14,7 +14,7 @@ class CaminhaoCliente extends \yii\db\ActiveRecord {
         return [
             [['caminhao_id', 'cliente_id', 'tipo_combustivel_id', 'valor_litro', 'valor_carrada', 'valor_frete', 'nota_fiscal', 'data', 'status'], 'required', 'on' => ['create', 'update'], 'message' => 'Campo obrigatório'],
             [['caminhao_id', 'cliente_id', 'tipo_combustivel_id', 'produto_negociacao_id', 'valor_carrada', 'status'], 'integer', 'on' => ['create', 'update']],
-            [['valor_carrada'], 'compare', 'compareValue' => (int) ValorSaida::find()->where(['produto_negociacao_id' => $this->produto_negociacao_id])->sum('valor'), 'operator' => '>=', 'on' => ['update'], 'message' => 'Limite de #' . (int) ValorSaida::find()->where(['produto_negociacao_id' => $this->produto_negociacao_id])->sum('valor') . ''],
+            [['valor_carrada'], 'compare', 'compareValue' => (double) ValorSaida::find()->where(['produto_negociacao_id' => $this->produto_negociacao_id])->sum('valor'), 'operator' => '>=', 'on' => ['update'], 'message' => 'Limite de #' . (double) ValorSaida::find()->where(['produto_negociacao_id' => $this->produto_negociacao_id])->sum('valor') . ''],
             [['valor_litro'], 'number', 'on' => ['create', 'update']],
             [['data'], 'safe', 'on' => ['create', 'update']],
             [['nota_fiscal', 'observacao'], 'string', 'max' => 500, 'on' => ['create', 'update']],
